Supprot file drop
Former-commit-id: 549492557aee10e122bdee660308149e40b2871a

+++ render.py
@@ -426,7 +426,8 @@ class Files(list):
         """Archive file then remove it.  """
 
         cls.archive(f)
-        os.remove(get_encoded(f))
+        if not os.path.isabs(f):
+            os.remove(get_encoded(f))
 
     @staticmethod
     def split_version(f):
